refactor: storage 파일 domain폴더로 이동
